functions file in every supposed dir
