+++ src/database/models.py
@@ -4,8 +4,9 @@ import requests
 from bs4 import BeautifulSoup
 from peewee import *
 
-from src.constants import MY_DATABASE, DATABASE_PATH, does_exist, determine_same_stop_names
-from src.functions import convert
+from src.constants import MY_DATABASE
+from src.database.functions import determine_same_stop_names
+from src.utils.functions import convert
 
 
 class Filter:
@@ -192,13 +193,3 @@ def fill_schedule(sources):
         ]).execute()
 
 
-def create_database(routes_list, fill_schedule_flag=False, db=MY_DATABASE, _filter=Filter()):
-    if not does_exist(DATABASE_PATH):
-        db.create_tables(DATABASE_TIMETABLES_LIST)
-    else:
-        print("=== database already exists! ===")
-        return
-
-    if fill_schedule_flag:
-        sources = gather_schedule_sources(routes_list, _filter)
-        fill_schedule(sources)
