+++ src/utils/functions.py
@@ -1,0 +1,5 @@
+import json
+
+
+def convert(data: dict) -> str:
+    return json.dumps(data, indent=2, separators=(',', ': '), default=str, ensure_ascii=False)
